Refactored code layout to be more reusable, primarily moving utilty functions into locations that make more sense (e.g., modelforge/curation/utils.py.py were moved to appropriate files under modelforge/utils/* ).  This included moving the figshare downloader to helper functions related to zenodo and general url processing (modelforge/utils/remote.py). Added zenodo downloader (similar to downloading figshare).
