+++ modelforge/curation/curation_baseclass.py
@@ -2,7 +2,68 @@ from typing import Optional
 from loguru import logger
 
 from abc import ABC, abstractmethod
-import modelforge.curation.units
+
+
+def dict_to_hdf5(file_name: str, data: list, series_info: dict, id_key: str) -> None:
+    """
+    Writes an hdf5 file from a list of dicts.
+
+    This will include units, if provided as attributes and also denote
+
+    Parameters
+    ----------
+    file_name: str, required
+        Name and path of hdf5 file to write.
+    data: list of dicts, required
+        List that contains dictionaries of properties for each molecule to write to file.
+    series_info: dict, required
+        Defines whether a piece of data containers a series of data associated with different conformers.
+        Dictionary keys match keys in dicts as part of data, where  "series" a
+    id_key: str, required
+        Name of the key in the dicts that uniquely describes each record.
+
+    Examples
+    --------
+    >>> series = {'name':False, 'atomic_numbers': False, 'n_configs': False, 'geometry': True, 'energy':True}
+    >>> dict_to_hdf5(file_name='qm9.hdf5', data=data, series_info=series, id_key='name')
+    """
+
+    import h5py
+    from tqdm import tqdm
+    import numpy as np
+    import pint
+
+    assert file_name.endswith(".hdf5")
+
+    dt = h5py.special_dtype(vlen=str)
+
+    with h5py.File(file_name, "w") as f:
+        for datapoint in tqdm(data):
+            try:
+                record_name = datapoint[id_key]
+            except Exception:
+                print(f"id_key {id_key} not found in the data.")
+            group = f.create_group(record_name)
+            for key, val in datapoint.items():
+                if key != id_key:
+                    if isinstance(val, pint.Quantity):
+                        val_m = val.m
+                        val_u = str(val.u)
+                    else:
+                        val_m = val
+                        val_u = None
+                    if isinstance(val_m, str):
+                        group.create_dataset(name=key, data=val_m, dtype=dt)
+                    elif isinstance(val_m, (float, int)):
+                        group.create_dataset(name=key, data=val_m)
+                    elif isinstance(val_m, np.ndarray):
+                        group.create_dataset(name=key, data=val_m, shape=val_m.shape)
+                    if not val_u is None:
+                        group[key].attrs["u"] = val_u
+                    if series_info[key]:
+                        group[key].attrs["series"] = True
+                    else:
+                        group[key].attrs["series"] = False
 
 
 class dataset_curation(ABC):
@@ -57,7 +118,7 @@ class dataset_curation(ABC):
         Creates an HDF5 file of the data at the path specified by output_file_path.
 
         """
-        from modelforge.curation.utils import dict_to_hdf5, mkdir
+        from modelforge.utils.misc import mkdir
 
         mkdir(self.output_file_dir)
 
@@ -104,12 +165,16 @@ class dataset_curation(ABC):
 
     @abstractmethod
     def _init_record_entries_series(self):
+        # For data efficiency, information for different conformers will be grouped together
+        # To make it clear to the dataset loader which pieces of information are common to all
+        # conformers, or which pieces encode the series, we will label each value.
         # The keys in this dictionary correspond to the label of the entries in each record.
-        # In this dictionary, the value indicates if the entry contains series data or just a single datapoint.
-        # If the entry has a value of "series", the "series" attribute in hdf5 file will be set to True (false if single)
+        # The value indicates if the entry contains series data (True) or a single common entry (False).
+        # If the entry has a value of True, the "series" attribute in hdf5 file will be set to True; False, if False.
         # This information will be used by the code to read in the datafile to know how to parse underlying records.
-        # While we could create separate records for every configuration, this vastly increases the time for generating
-        # and reading hdf5 files.
+        # Example where the name and atomic numbers fields contain only a single common entry, but geometry
+        # and energy are a series for each conformer:
+        # self._record_entries_series = {'name':False, 'atomic_numbers': False, 'geometry': True, 'energy':True}
 
         self._record_entries_series = {}
