+++ modelforge/dataset/transformation.py
@@ -1,6 +1,6 @@
 import numpy as np
 
-from .utils import pad_molecules, pad_to_max_length
+from modelforge.dataset.utils import pad_molecules, pad_to_max_length
 
 default_transformation = {
     "geometry": pad_molecules,
