+++ modelforge/tests/test_remote.py
@@ -1,7 +1,15 @@
 import sys
 import pytest
 import modelforge
-from modelforge.utils.zenodo import *
+import os
+
+from modelforge.utils.remote import *
+
+
+@pytest.fixture(scope="session")
+def prep_temp_dir(tmp_path_factory):
+    fn = tmp_path_factory.mktemp("remote_test")
+    return fn
 
 
 def test_is_url():
@@ -16,6 +24,26 @@ def test_is_url():
         is_url(query="https://zenodo.org/record/3588339", hostname="choderalab.org")
         == False
     )
+
+
+def test_download_from_figshare(prep_temp_dir):
+    url = "https://figshare.com/ndownloader/files/22247589"
+    name = download_from_figshare(
+        url=url,
+        output_path=str(prep_temp_dir),
+        force_download=True,
+    )
+
+    file_name_path = str(prep_temp_dir) + f"/{name}"
+    assert os.path.isfile(file_name_path)
+
+    with pytest.raises(Exception):
+        url = "https://choderalab.com/ndownloader/files/22247589"
+        name = download_from_figshare(
+            url=url,
+            output_path=str(prep_temp_dir),
+            force_download=True,
+        )
 
 
 def test_record_id_parse():
@@ -33,28 +61,34 @@ def test_record_id_parse():
 
 def test_zenodo_fetch():
     # qm9 dataset
-    files = hdf5_from_zenodo(record="10.5281/zenodo.3588339")
+    files = datafiles_from_zenodo(
+        record="10.5281/zenodo.3588339", file_extension="hdf5.gz"
+    )
     assert len(files) == 1
     assert (
         files[0]
         == "https://zenodo.org/api/files/0a55a53a-69c3-4cd8-8ab8-d031ca0d6853/155.hdf5.gz"
     )
 
-    files = hdf5_from_zenodo(record="https://dx.doi.org/10.5281/zenodo.3588339")
+    files = datafiles_from_zenodo(
+        record="https://dx.doi.org/10.5281/zenodo.3588339", file_extension="hdf5.gz"
+    )
     assert len(files) == 1
     assert (
         files[0]
         == "https://zenodo.org/api/files/0a55a53a-69c3-4cd8-8ab8-d031ca0d6853/155.hdf5.gz"
     )
 
-    files = hdf5_from_zenodo(record="https://zenodo.org/record/3588339")
+    files = datafiles_from_zenodo(
+        record="https://zenodo.org/record/3588339", file_extension="hdf5.gz"
+    )
     assert len(files) == 1
     assert (
         files[0]
         == "https://zenodo.org/api/files/0a55a53a-69c3-4cd8-8ab8-d031ca0d6853/155.hdf5.gz"
     )
 
-    files = hdf5_from_zenodo(record="3588339")
+    files = datafiles_from_zenodo(record="3588339", file_extension="hdf5.gz")
     assert len(files) == 1
     assert (
         files[0]
@@ -62,7 +96,7 @@ def test_zenodo_fetch():
     )
 
     # comp6 dataset
-    files = hdf5_from_zenodo("10.5281/zenodo.3588368")
+    files = datafiles_from_zenodo("10.5281/zenodo.3588368", file_extension="hdf5.gz")
     assert len(files) == 6
     assert any("205.hdf5.gz" in file for file in files)
     assert any("207.hdf5.gz" in file for file in files)
@@ -83,8 +117,10 @@ def test_fetch_record_id():
         fetch_url_from_doi(doi="10.5281/zenodo.3588339", timeout=0.0000000000001)
 
 
-def test_fetch_data():
-    files = get_zenodo_datafiles(record_id="3588339", file_extension=".hdf5.gz")
+def test_fetch_datafiles():
+    files = get_zenodo_datafiles(
+        record_id="https://zenodo.org/record/3588339", file_extension=".hdf5.gz"
+    )
 
     assert len(files) == 1
     assert (
@@ -112,4 +148,24 @@ def test_fetch_data():
     with pytest.raises(Exception):
         get_zenodo_datafiles(
             record_id="3588339", timeout=0.000000000001, file_extension=".hdf5.gz"
+        )
+
+
+def test_download_from_zenodo(prep_temp_dir):
+    url = "https://zenodo.org/record/3401581/files/PTC-CMC/atools_ml-v0.1.zip"
+    name = download_from_zenodo(
+        url=url,
+        output_path=str(prep_temp_dir),
+        force_download=True,
+    )
+
+    file_name_path = str(prep_temp_dir) + f"/{name}"
+    assert os.path.isfile(file_name_path)
+
+    with pytest.raises(Exception):
+        url = "https://choderalab.com/22247589"
+        name = download_from_zenodo(
+            url=url,
+            output_path=str(prep_temp_dir),
+            force_download=True,
         )
